+++ src/components/pwc-map-controls/components/pwc-map-text-control/pwc-map-text-control.css
@@ -1,0 +1,2 @@
+pwc-map-text-control {
+}
