--- src/components/pwc-map-controls/components/pwc-text-control/pwc-text-control.css
@@ -1,2 +1,0 @@
-pwc-text-control {
-}
